Some fixes Now using usefixtures decorator; Added confest.py files; Removed some repeats.
