+++ apps/profiles/tests/followers/tests_accept.py
@@ -1,23 +1,20 @@
-import os
-import django
-
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "config.settings")
-django.setup()
-
+import pytest
 from rest_framework.reverse import reverse
 from rest_framework.exceptions import ErrorDetail
 
-from apps.profiles.tests.fixtures import *
+from apps.profiles.tests.utils import get_tokens
 from apps.profiles.tests.followers.constants import ACCEPT_URL
 from apps.profiles.models.followers import Follower
 
 
+@pytest.mark.usefixtures(
+    "follower_user_pending_private",
+)
 @pytest.mark.django_db
 def test_accept_pending_private(
         api_client,
         user_private,
         user_2_private,
-        follower_user_pending_private,
 ):
     token = get_tokens(user_2_private)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
@@ -31,12 +28,14 @@ def test_accept_pending_private(
     assert response.data.get('status') == Follower.Status.ACCEPTED
 
 
+@pytest.mark.usefixtures(
+    "follower_user_declined_private",
+)
 @pytest.mark.django_db
 def test_accept_declined_private(
         api_client,
         user_private,
         user_2_private,
-        follower_user_declined_private,
 ):
     token = get_tokens(user_2_private)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
@@ -99,12 +98,14 @@ def test_accept_no_creds(
     }
 
 
+@pytest.mark.usefixtures(
+    "follower_user_accepted",
+)
 @pytest.mark.django_db
 def test_accept_accepted(
         api_client,
         user,
         user_2,
-        follower_user_accepted,
 ):
     token = get_tokens(user_2)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
