+++ apps/profiles/tests/followers/tests_follow.py
@@ -1,13 +1,8 @@
-import os
-import django
-
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "config.settings")
-django.setup()
-
+import pytest
 from rest_framework.reverse import reverse
 from rest_framework.exceptions import ErrorDetail
 
-from apps.profiles.tests.fixtures import *
+from apps.profiles.tests.utils import *
 from apps.profiles.tests.followers.constants import FOLLOW_URL
 from apps.profiles.models.followers import Follower
 
@@ -97,12 +92,14 @@ def test_follow_no_creds(
     }
 
 
+@pytest.mark.usefixtures(
+    "follower_user_accepted",
+)
 @pytest.mark.django_db
 def test_follow_accepted(
         api_client,
         user,
         user_2,
-        follower_user_accepted,
 ):
     token = get_tokens(user)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
@@ -113,12 +110,14 @@ def test_follow_accepted(
     assert response.data == {'detail': 'Already following'}
 
 
+@pytest.mark.usefixtures(
+    "follower_user_pending_private",
+)
 @pytest.mark.django_db
 def test_follow_pending_private(
         api_client,
         user_private,
         user_2_private,
-        follower_user_pending_private,
 ):
     token = get_tokens(user_private)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
@@ -129,12 +128,14 @@ def test_follow_pending_private(
     assert response.data == {'detail': 'Follow request already sent'}
 
 
+@pytest.mark.usefixtures(
+    "follower_user_declined_private",
+)
 @pytest.mark.django_db
 def test_follow_declined(
         api_client,
         user_private,
         user_2_private,
-        follower_user_declined_private,
 ):
     token = get_tokens(user_private)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
