+++ apps/profiles/tests/followers/tests_list_followers.py
@@ -1,13 +1,8 @@
-import os
-import django
-
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "config.settings")
-django.setup()
-
+import pytest
 from rest_framework.reverse import reverse
 from rest_framework.exceptions import ErrorDetail
 
-from apps.profiles.tests.fixtures import *
+from apps.profiles.tests.utils import *
 from apps.profiles.tests.followers.constants import LIST_FOLLOWERS_URL
 from apps.profiles.models.followers import Follower
 
@@ -27,12 +22,14 @@ def test_list_followers_without_following(
     assert response.data == []
 
 
+@pytest.mark.usefixtures(
+    "follower_user_accepted",
+)
 @pytest.mark.django_db
 def test_list_followers_accepted(
         api_client,
         user,
         user_2,
-        follower_user_accepted,
 ):
     token = get_tokens(user)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
@@ -46,12 +43,14 @@ def test_list_followers_accepted(
     assert response.data[0].get("status") == Follower.Status.ACCEPTED
 
 
+@pytest.mark.usefixtures(
+    "follower_user_accepted_private",
+)
 @pytest.mark.django_db
 def test_list_followers_accepted_private(
         api_client,
         user_private,
         user_2_private,
-        follower_user_accepted_private,
 ):
     token = get_tokens(user_private)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
@@ -85,12 +84,14 @@ def test_list_followers_without_following_private(
     }
 
 
+@pytest.mark.usefixtures(
+    "follower_user_pending_private",
+)
 @pytest.mark.django_db
 def test_list_followers_pending_private(
         api_client,
         user,
         user_2_private,
-        follower_user_pending_private,
 ):
     token = get_tokens(user)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
@@ -106,12 +107,14 @@ def test_list_followers_pending_private(
     }
 
 
+@pytest.mark.usefixtures(
+    "follower_user_declined_private",
+)
 @pytest.mark.django_db
 def test_list_followers_declined_private(
         api_client,
         user,
         user_2_private,
-        follower_user_declined_private,
 ):
     token = get_tokens(user)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
@@ -127,12 +130,14 @@ def test_list_followers_declined_private(
     }
 
 
+@pytest.mark.usefixtures(
+    "follower_user_accepted",
+)
 @pytest.mark.django_db
 def test_list_followers_current_user_accepted(
         api_client,
         user,
         user_2,
-        follower_user_accepted,
 ):
     token = get_tokens(user_2)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
@@ -146,12 +151,14 @@ def test_list_followers_current_user_accepted(
     assert response.data[0].get("status") == Follower.Status.ACCEPTED
 
 
+@pytest.mark.usefixtures(
+    "follower_user_pending_private",
+)
 @pytest.mark.django_db
 def test_list_followers_current_user_pending_private(
         api_client,
         user,
         user_2_private,
-        follower_user_pending_private,
 ):
     token = get_tokens(user_2_private)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
@@ -162,12 +169,14 @@ def test_list_followers_current_user_pending_private(
     assert response.data == []
 
 
+@pytest.mark.usefixtures(
+    "follower_user_declined_private",
+)
 @pytest.mark.django_db
 def test_list_followers_current_user_declined_private(
         api_client,
         user,
         user_2_private,
-        follower_user_declined_private,
 ):
     token = get_tokens(user_2_private)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
