+++ apps/profiles/tests/followers/tests_list_following.py
@@ -1,23 +1,20 @@
-import os
-import django
-
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "config.settings")
-django.setup()
-
+import pytest
 from rest_framework.reverse import reverse
 from rest_framework.exceptions import ErrorDetail
 
-from apps.profiles.tests.fixtures import *
+from apps.profiles.tests.utils import *
 from apps.profiles.tests.followers.constants import LIST_FOLLOWING_URL
 from apps.profiles.models.followers import Follower
 
 
+@pytest.mark.usefixtures(
+    "follower_user_2_accepted",
+)
 @pytest.mark.django_db
 def test_list_following_without_following(
         api_client,
         user,
         user_2,
-        follower_user_2_accepted,
 ):
     token = get_tokens(user)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
@@ -31,13 +28,15 @@ def test_list_following_without_following(
     assert response.data[0].get("status") == Follower.Status.ACCEPTED
 
 
+@pytest.mark.usefixtures(
+    "follower_user_accepted",
+    "follower_user_2_accepted",
+)
 @pytest.mark.django_db
 def test_list_following_accepted(
         api_client,
         user,
         user_2,
-        follower_user_accepted,
-        follower_user_2_accepted,
 ):
     token = get_tokens(user)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
@@ -51,13 +50,15 @@ def test_list_following_accepted(
     assert response.data[0].get("status") == Follower.Status.ACCEPTED
 
 
+@pytest.mark.usefixtures(
+    "follower_user_accepted_private",
+    "follower_user_2_accepted_private",
+)
 @pytest.mark.django_db
 def test_list_following_accepted_private(
         api_client,
         user_private,
         user_2_private,
-        follower_user_accepted_private,
-        follower_user_2_accepted_private,
 ):
     token = get_tokens(user_private)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
@@ -91,13 +92,15 @@ def test_list_following_without_following_private(
     }
 
 
+@pytest.mark.usefixtures(
+    "follower_user_pending_private",
+    "follower_user_2_accepted_private",
+)
 @pytest.mark.django_db
 def test_list_following_pending_private(
         api_client,
         user_private,
         user_2_private,
-        follower_user_pending_private,
-        follower_user_2_accepted_private,
 ):
     token = get_tokens(user_private)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
@@ -113,13 +116,15 @@ def test_list_following_pending_private(
     }
 
 
+@pytest.mark.usefixtures(
+    "follower_user_declined_private",
+    "follower_user_2_accepted_private",
+)
 @pytest.mark.django_db
 def test_list_following_declined_private(
         api_client,
         user_private,
         user_2_private,
-        follower_user_declined_private,
-        follower_user_2_accepted_private,
 ):
     token = get_tokens(user_private)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
@@ -135,12 +140,14 @@ def test_list_following_declined_private(
     }
 
 
+@pytest.mark.usefixtures(
+    "follower_user_2_accepted",
+)
 @pytest.mark.django_db
 def test_list_following_current_user_accepted(
         api_client,
         user,
         user_2,
-        follower_user_2_accepted,
 ):
     token = get_tokens(user_2)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
@@ -154,12 +161,14 @@ def test_list_following_current_user_accepted(
     assert response.data[0].get("status") == Follower.Status.ACCEPTED
 
 
+@pytest.mark.usefixtures(
+    "follower_user_2_pending_private",
+)
 @pytest.mark.django_db
 def test_list_following_current_user_pending_private(
         api_client,
         user_private,
         user_2,
-        follower_user_2_pending_private,
 ):
     token = get_tokens(user_2)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
@@ -170,12 +179,14 @@ def test_list_following_current_user_pending_private(
     assert response.data == []
 
 
+@pytest.mark.usefixtures(
+    "follower_user_2_declined_private",
+)
 @pytest.mark.django_db
 def test_list_following_current_user_declined_private(
         api_client,
         user_private,
         user_2,
-        follower_user_2_declined_private,
 ):
     token = get_tokens(user_2)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
